101-mul

+++ 0x0C-more_malloc_free/101-mul.c
@@ -1,4 +1,7 @@
 #include "main.h"
+#include <stdlib.h>
+#include <stdio.h>
+#include <ctype.h>
 
 /**
  * main - multiplies two positive numbers
@@ -27,7 +30,7 @@ int i, j;
 			exit(98); }
 		}
 	}
-	mul = atoi(argv[1]) * atoi(argv[2]);
+	mul = atol(argv[1]) * atol(argv[2]);
 	printf("%lu\n", mul);
 
 return (0);
